refactor CMakeLists into subdirs and create include dir for headers
