+++ src/calculator.cpp
@@ -1,4 +1,4 @@
-#include "calculator.hpp"
+#include "calculator/calculator.hpp"
 #include <iostream>
 
 Calculator::Calculator(Logger logger, Notifier notifier) : logger(logger), notifier(notifier) {}
